fix(components/atom/input): input mask fully hooked
